+++ src/Twilio/Rest/Messaging/V1/Service/AlphaSenderResource.cs
@@ -423,7 +423,7 @@ namespace Twilio.Rest.Messaging.V1.Service
         /// An array of values that describe whether the number can receive calls or messages
         /// </summary>
         [JsonProperty("capabilities")]
-        public List<object> Capabilities { get; private set; }
+        public List<string> Capabilities { get; private set; }
         /// <summary>
         /// The absolute URL of the AlphaSender resource
         /// </summary>